+++ needle/modules/dynamic/memory/heap_dump.py
@@ -59,8 +59,7 @@ class Module(BaseModule):
         # Check if we have dumps
         self.printer.verbose("Checking if we have dumps...")
         file_list = self.device.remote_op.dir_list(dir_dumps, recursive=True)
-        failure = filter(lambda x: 'total 0' in x, file_list)
-        if failure:
+        if not file_list:
             self.printer.error('It was not possible to attach to the process (known issue in iOS9. A Fix is coming soon)')
             return